requirements.txt
PyQt6
openai>=1.0.0
sounddevice
numpy
scipy
keyboard
pyperclip
python-dotenv
httpx<0.28.0
